+++ bonfire/libraries/installer_lib.php
@@ -567,7 +567,7 @@ class Installer_lib {
 	{
 		$mod_versions = array();
 
-		$modules = module_files(null, 'migrations');
+		$modules = Modules::files(null, 'migrations');
 
 		if ($modules === false)
 		{